Worked on the portal mechanics. Also turned the portal into a prefab for easy reuse.
Needs some cleanup in the code, but the portals now behave as I intended.
Rather than having to go to the portal door mesh, we can toggle their state and their level to teleport from the root object.

Test portal to the left is always on and teleports to Tutorial.
Portal/The exit of the level is already on and teleports to Navigation prototype.
That level's portal is inactive and can be activated by pressing the button.